requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
requests==2.32.3
openai==1.65.2
transformers==4.49.0
torch==2.6.0
tqdm==4.67.1
tiktoken==0.9.0
python-dotenv==1.0.1
pdfplumber==0.11.6
